requirements.txt
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
boto3==1.42.89
botocore==1.42.89
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
Django==6.0.3
django-unfold==0.87.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
google-auth==2.49.1
h11==0.16.0
httptools==0.7.1
idna==3.11
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.4.3
pgvector==0.4.2
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
whitenoise==6.12.0